# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unlimited-OCR Inference Server",
    "version": "0.0.0"
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Liveness check",
        "security": [],
        "responses": {
          "200": {
            "description": "Process is alive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness check",
        "security": [],
        "responses": {
          "200": {
            "description": "At least one worker is ready"
          },
          "503": {
            "description": "No model worker is ready"
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Runtime metrics snapshot",
        "responses": {
          "200": {
            "description": "Metrics snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/infer": {
      "post": {
        "summary": "Queue inference from a multipart image or PDF upload",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadInferenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job or PDF page jobs queued",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/QueueResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PdfQueueResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/infer/path": {
      "post": {
        "summary": "Queue inference from a server-side image or PDF path",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalPathRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job or PDF page jobs queued",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/QueueResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PdfQueueResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "summary": "Fetch a job record",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "API key is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Request rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Service unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "QueueResponse": {
        "type": "object",
        "required": [
          "id",
          "status",
          "status_url"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "status_url": {
            "type": "string"
          }
        }
      },
      "PdfQueueResponse": {
        "type": "object",
        "required": [
          "kind",
          "page_count",
          "jobs"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "const": "pdf"
          },
          "page_count": {
            "type": "integer"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueResponse"
            }
          }
        }
      },
      "LocalPathRequest": {
        "type": "object",
        "required": [
          "image_path"
        ],
        "properties": {
          "image_path": {
            "type": "string",
            "description": "Server-side image or PDF path"
          },
          "text_input": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional Unlimited-OCR prompt override. Defaults to <image>Free OCR."
          },
          "webhook_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
      "UploadInferenceRequest": {
        "type": "object",
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Image or PDF file"
          },
          "text_input": {
            "type": "string",
            "description": "Optional Unlimited-OCR prompt override. Defaults to <image>Free OCR."
          },
          "webhook_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "WorkerHealth": {
        "type": "object",
        "required": [
          "expected",
          "ready",
          "failed"
        ],
        "properties": {
          "expected": {
            "type": "integer"
          },
          "ready": {
            "type": "integer"
          },
          "failed": {
            "type": "integer"
          }
        }
      },
      "HealthResponse": {
        "type": "object"
      },
      "MetricsResponse": {
        "type": "object"
      },
      "JobRecord": {
        "type": "object"
      }
    }
  }
}